Move @named_data into main ddt.py module. Leave NamedData types in a named_data_types.py.
